backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
motor
python-multipart
livekit-api
certifi
huggingface_hub
pypdf
pillow
